First tested (leaking) MS queue implementation
Now the MS queue passes some initial unit tests, and also passes
them in Miri. The difference from last commit is using the CAS
for pointers rather than the normal safe one. I don't really know
why, but Miri complained when using the normal one.

The remaining issue is that the queue now leaks memory, as it never
frees the remaining nodes in the queue when dropped.